+++ examples/caching_requestor.py
@@ -9,6 +9,7 @@ Demonstrates the use of custom sessions with :class:`.Requestor`. It's an adapta
 
 import os
 import sys
+from typing import ClassVar
 
 import requests
 
@@ -23,17 +24,17 @@ class CachingSession(requests.Session):
 
     """
 
-    get_cache = {}
+    get_cache: ClassVar = {}
 
     def request(self, method, url, params=None, **kwargs):
         """Perform a request, or return a cached response if available."""
         params_key = tuple(params.items()) if params else ()
         if method.upper() == "GET" and (url, params_key) in self.get_cache:
             print("Returning cached response for:", method, url, params)
-            return self.get_cache[(url, params_key)]
+            return self.get_cache[url, params_key]
         result = super().request(method, url, params, **kwargs)
         if method.upper() == "GET":
-            self.get_cache[(url, params_key)] = result
+            self.get_cache[url, params_key] = result
             print("Adding entry to the cache:", method, url, params)
         return result
 
@@ -44,7 +45,7 @@ def main():
         print(f"Usage: {sys.argv[0]} USERNAME")
         return 1
 
-    caching_requestor = prawcore.Requestor("prawcore_device_id_auth_example", session=CachingSession())
+    caching_requestor = prawcore.Requestor(user_agent="prawcore_device_id_auth_example", session=CachingSession())
     authenticator = prawcore.TrustedAuthenticator(
         caching_requestor,
         os.environ["PRAWCORE_CLIENT_ID"],
@@ -55,10 +56,10 @@ def main():
 
     user = sys.argv[1]
     with prawcore.session(authorizer) as session:
-        data1 = session.request("GET", f"/api/v1/user/{user}/trophies")
+        data1 = session.request(method="GET", path=f"/api/v1/user/{user}/trophies")
 
     with prawcore.session(authorizer) as session:
-        data2 = session.request("GET", f"/api/v1/user/{user}/trophies")
+        data2 = session.request(method="GET", path=f"/api/v1/user/{user}/trophies")
 
     for trophy in data1["data"]["trophies"]:
         description = trophy["data"]["description"]