+++ examples/device_id_auth_trophies.py
@@ -19,7 +19,7 @@ def main():
         return 1
 
     authenticator = prawcore.UntrustedAuthenticator(
-        prawcore.Requestor("prawcore_device_id_auth_example"),
+        prawcore.Requestor(user_agent="prawcore_device_id_auth_example"),
         os.environ["PRAWCORE_CLIENT_ID"],
     )
     authorizer = prawcore.DeviceIDAuthorizer(authenticator)
@@ -27,7 +27,7 @@ def main():
 
     user = sys.argv[1]
     with prawcore.session(authorizer) as session:
-        data = session.request("GET", f"/api/v1/user/{user}/trophies")
+        data = session.request(method="GET", path=f"/api/v1/user/{user}/trophies")
 
     for trophy in data["data"]["trophies"]:
         description = trophy["data"]["description"]